refactor and prepare api_page_helper for forecast_weather api
