+++ app/helpers/weather_helper.rb
@@ -1,0 +1,24 @@
+module WeatherHelper
+  # from here on get all data out of json(curr_weather_data)
+  def get_weather_data(curr_weather_data)
+    weather_pic(curr_weather_data)
+    @temp = ((curr_weather_data['main']['temp']).to_f - 273.15).round(1)
+    @city = curr_weather_data['name']
+    cloud_view(curr_weather_data)
+    @time = Time.now # change this part!!
+  end
+
+  def cloud_view(curr_weather_data)
+    all_clouds_number = curr_weather_data['clouds']['all']
+    if all_clouds_number == 0
+      @sky_view = 'sky is clear'
+    else
+      @sky_view = 'sky is not clear'
+    end
+  end
+
+  def weather_pic(curr_weather_data)
+    # add diff weather cases to the pisc from openweathermap.org
+    @weather_pic = 'http://openweathermap.org/img/w/01d.png'
+  end
+end
