+++ app/views/api_page/index.html.erb
@@ -5,9 +5,7 @@
 </div></br>
 
 <hr>
-
 <%= render 'curr_weather_api' %>
-
 <hr>
-
 <%= render 'forecast_weather_api' %>
+<hr>
